test(chat): align testing strategy with no-mock policy and expand coverage
- AGENTS.md: remove guidance to mock Supabase in unit tests; clarify
  that service tests belong in __integration__ and store tests mock
  the service layer only
- services/__tests__/chat.service.test.ts: remove all Supabase-mocked
  tests (covered by integration tests); keep pure Zod schema tests
- store/__tests__/chat.store.test.ts: expand from 2 to 15 tests,
  covering getAllChatRooms, getAllMessages, loadMoreMessages,
  openChatRoom, closeChatRoom, and addMessage
- __integration__/chat.test.ts: add US2-6 (since param filtering) and
  US2-7 (markChatRead updates last_read_at and clears has_unread)


fix(tests): move misplaced integration tests to __integration__/ and fix rate limiting

services/__tests__/trip.service.test.ts and announcements.service.test.ts
were integration tests (calling createTestUser and real Supabase) placed in
the wrong folder. Running in parallel with other integration tests caused
Supabase auth rate limiting, breaking the pre-push hook.

- Move trip integration coverage to __integration__/trip-roles.test.ts
  (unique updateParticipantRole actor-permission tests not in trip-participant.test.ts)
- Move announcement integration coverage to __integration__/announcements.test.ts
- Keep only pure Zod schema tests in the services/__tests__/ files
- Add maxWorkers: 4 to jest config to cap test parallelism


made changes to comply with latest layout.
