+++ package.json
@@ -64,6 +64,7 @@
   },
   "jest": {
     "preset": "jest-expo",
+    "maxWorkers": 4,
     "setupFiles": [
       "./jest.setup.ts"
     ],
